add new field to post models and update handlers

+++ models/user.go
@@ -3,7 +3,7 @@ package models
 import "gorm.io/gorm"
 
 type User struct {
-	ID        int     `gorm:"primaryKey" json:"id"`
+	ID        int64   `gorm:"primaryKey" json:"id"`
 	FirstName string  `gorm:"size:20" json:"first_name"`
 	LastName  string  `gorm:"size:20" json:"last_name"`
 	Username  string  `gorm:"size:20;unique" json:"username"`
@@ -16,7 +16,7 @@ type UserModel interface {
 	// CreateUser function to create new user.
 	CreateUser(user *User) error
 	// GetUserByID function to get user by id
-	GetUserByID(id int) (User, error)
+	GetUserByID(id int64) (User, error)
 	// GetUserByUsername function to get user by username.
 	GetUserByUsername(username string) (User, error)
 }
@@ -33,7 +33,7 @@ func (u userModel) CreateUser(user *User) error {
 	return u.db.Model(&User{}).Create(user).Error
 }
 
-func (u userModel) GetUserByID(id int) (User, error) {
+func (u userModel) GetUserByID(id int64) (User, error) {
 	var user User
 	err := u.db.Model(&User{}).Where("id = ?", id).First(&user).Error
 	return user, err
